move BancoHelperImpl::convertAmountToItems to ItemUtil::convertAmountToItems delete BancoHelperImpl::removeFromInventory
