+++ common/src/main/java/ovh/mythmc/banco/common/impl/BancoHelperImpl.java
@@ -9,15 +9,11 @@ import ovh.mythmc.banco.api.economy.BancoHelper;
 import ovh.mythmc.banco.api.economy.BancoHelperSupplier;
 import ovh.mythmc.banco.api.inventories.BancoInventory;
 import ovh.mythmc.banco.api.items.BancoItem;
-import ovh.mythmc.banco.api.accounts.Account;
 import ovh.mythmc.banco.common.inventories.EnderChestInventoryImpl;
 import ovh.mythmc.banco.common.inventories.PlayerInventoryImpl;
 import ovh.mythmc.banco.common.util.ItemUtil;
 
 import java.math.BigDecimal;
-import java.math.RoundingMode;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.UUID;
 
 public class BancoHelperImpl implements BancoHelper {
@@ -27,11 +23,13 @@ public class BancoHelperImpl implements BancoHelper {
 
     public BancoHelperImpl() {
         BancoHelperSupplier.set(this);
-
-        // Banco inventories
         this.playerInventory = new PlayerInventoryImpl();
         this.enderChestInventory = new EnderChestInventoryImpl();
-        Banco.get().getInventoryManager().registerInventory(playerInventory, enderChestInventory);
+
+        // Register banco inventories
+        Banco.get().getInventoryManager().registerInventory(playerInventory);
+        if (Banco.get().getSettings().get().getCurrency().isCountEnderChest())
+            Banco.get().getInventoryManager().registerInventory(enderChestInventory);
     }
 
     // Todo: unnecessary?
@@ -41,9 +39,10 @@ public class BancoHelperImpl implements BancoHelper {
         if (player == null)
             return BigDecimal.valueOf(0);
 
-        return playerInventory.add(uuid, amount);
+        return amount.subtract(playerInventory.add(uuid, amount));
     }
 
+    // Todo: unnecessary?
     // Todo: look for less valuable items first
     @Override
     public BigDecimal remove(UUID uuid, BigDecimal amount) {
@@ -51,39 +50,9 @@ public class BancoHelperImpl implements BancoHelper {
         if (player == null)
             return BigDecimal.valueOf(0);
 
-        if (Banco.get().getSettings().get().getCurrency().isCountEnderChest()) {
-            BigDecimal remainingAmount = removeFromInventory(player.getEnderChest().getContents(), uuid, amount);
-
-            if (remainingAmount.compareTo(BigDecimal.valueOf(0)) > 0)
-                return removeFromInventory(player.getInventory().getContents(), uuid, remainingAmount);
-        }
-
-        return removeFromInventory(player.getInventory().getContents(), uuid, amount);
-    }
-
-    private BigDecimal removeFromInventory(ItemStack[] inventory, UUID uuid, BigDecimal amount) {
-        for (ItemStack item : inventory) {
-            if (item == null) continue;
-            if (amount.compareTo(BigDecimal.valueOf(0)) < 0.01) continue;
-
-            BigDecimal value = BigDecimal.valueOf(0);
-
-            BancoItem bancoItem = ItemUtil.getBancoItem(item);
-            if (bancoItem != null)
-                value = value.add(Banco.get().getEconomyManager().value(bancoItem, item.getAmount()));
-
-            if (value.compareTo(BigDecimal.valueOf(0)) > 0) {
-                item.setAmount(0);
-                BigDecimal added = BigDecimal.valueOf(0);
-                if (value.compareTo(amount) > 0) {
-                    added = value.subtract(amount);
-                    Account account = Banco.get().getAccountManager().get(uuid);
-                    if (account != null)
-                        Banco.get().getAccountManager().set(account, account.amount().add(added));
-                }
-
-                amount = amount.subtract(value.subtract(added));
-            }
+        for (BancoInventory<Inventory> inventory : Banco.get().getInventoryManager().get()) {
+            if (amount.compareTo(BigDecimal.valueOf(0)) > 0)
+                amount = inventory.remove(uuid, amount);
         }
 
         return amount;
@@ -109,25 +78,6 @@ public class BancoHelperImpl implements BancoHelper {
         }
 
         return value;
-    }
-
-    public static List<ItemStack> convertAmountToItems(BigDecimal amount) {
-        List<ItemStack> items = new ArrayList<>();
-
-        for (BancoItem bancoItem : Banco.get().getEconomyManager().get().reversed()) {
-            if(bancoItem.value().compareTo(amount) > 0)
-                continue;
-
-            int itemAmount = (amount.divide(bancoItem.value(), RoundingMode.FLOOR)).intValue();
-
-            if (itemAmount > 0) {
-                items.add(ItemUtil.getItemStack(bancoItem, itemAmount));
-
-                amount = amount.subtract(Banco.get().getEconomyManager().value(bancoItem, itemAmount));
-            }
-        }
-
-        return items;
     }
 
 }
